rste3-phll_study/requirements.txt
e3nn==0.5.1
hydra-core==1.3.2
matplotlib==3.9.2
numpy==2.1.3
Ofpp==0.12
Ofpp==0.12
omegaconf==2.3.0
PyFoam==2023.7
PyFoam==2023.7
python-dotenv==1.0.1
pytorch_lightning==1.5.0
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.5.2
torch==2.1.0+cu118
tqdm==4.66.4
utils==1.0.2
